Use inttypes PRI* macros throughout the kernel

+++ kos/src/kernel/core/fs/node.c
@@ -4121,8 +4121,8 @@ check_result_for_deletion:
 		assertf((result->i_filemode & S_IFMT) ==
 		        (DTTOIF(parent_directory_entry->de_type) & S_IFMT),
 		        "%u != %u",
-		        IFTODT(result->i_filemode & S_IFMT),
-		        parent_directory_entry->de_type);
+		        (unsigned int)IFTODT(result->i_filemode & S_IFMT),
+		        (unsigned int)parent_directory_entry->de_type);
 		if unlikely(ATOMIC_READ(result->i_flags) & INODE_FDELETED) {
 			/* Might happen due to race conditions. (start over...) */
 			superblock_nodeslock_endwrite(self);
